+++ paimon-common/src/main/java/org/apache/paimon/fileindex/FileIndexer.java
@@ -18,11 +18,11 @@
 
 package org.apache.paimon.fileindex;
 
-import org.apache.paimon.fileindex.bloomfilter.BloomFilter;
+import org.apache.paimon.fileindex.bloomfilter.BloomFilterFileIndex;
 import org.apache.paimon.options.Options;
 import org.apache.paimon.types.DataType;
 
-import static org.apache.paimon.fileindex.bloomfilter.BloomFilter.BLOOM_FILTER;
+import static org.apache.paimon.fileindex.bloomfilter.BloomFilterFileIndex.BLOOM_FILTER;
 
 /** File index interface. To build a file index. */
 public interface FileIndexer {
@@ -34,7 +34,7 @@ public interface FileIndexer {
     static FileIndexer create(String type, DataType dataType, Options options) {
         switch (type) {
             case BLOOM_FILTER:
-                return new BloomFilter(dataType, options);
+                return new BloomFilterFileIndex(dataType, options);
             default:
                 throw new RuntimeException("Doesn't support filter type: " + type);
         }
